move district cube to admin pages and remove components folder
