+++ frontend/lib/pages/admin/district_cube.dart
